==> 4. OES Data Clean (2).ipynb <==
## Purpose
Take the occupation employment statistics and map them onto the zip code level using principal building activity (PCA) code from CBECS. The end result is the number of workers in a given zip code in a given commercial building type.

import pandas as pd
import sqlite3
import numpy as np

Read in employment data

# ------------------------------------------- #
con = sqlite3.connect('Datasets/oes_may12_demo.db')
# ------------------------------------------- #

Get occupation data at the "major" occupation level. Here we are also going to divy up the total number employees amongst zip code. We will do it such that all zip codes within a given MSA will be given an even number of employees

df = pd.read_sql('select * from data where OCC_GROUP = \'major\'',con,index_col = 'index')
df_zip_count = df.groupby(['MSA No.','OCC_CODE'])[['id']].count()
df_zip_count.rename(columns = {'id':'nzip'},inplace=True)
df = df.join(df_zip_count,on=['MSA No.','OCC_CODE'])
df['EMP'] = df['TOT_EMP']/df['nzip']

df.shape

Get occupation data at the "total" occupation level. Here we are also going to divy up the total number employees amongst zip code. We will do it such that all zip codes within a given MSA will be given an even number of employees

df_tot = pd.read_sql('select * from data where OCC_GROUP = \'total\'',con,index_col = 'index')
df_tot_zip_count = df_tot.groupby(['MSA No.','OCC_CODE'])[['id']].count()
df_tot_zip_count.rename(columns = {'id':'nzip'},inplace=True)
df_tot = df_tot.join(df_tot_zip_count,on=['MSA No.','OCC_CODE'])
df_tot['EMP'] = df_tot['TOT_EMP']/df_tot['nzip']

Use this field for distributing employees amongst PBA types

field = 'EMP'

Read in dataframe with my mappings of job classifcations onto PBA

job = pd.read_excel('Datasets/job_class.xlsx',index_col=0) # Job class can be mapped with certain building types

Read in csv file with description of PBA types

pba_id = pd.read_csv('Datasets/pba_id.csv',index_col=0)

Create an empty truth table with job types as rows and PBAs as columns. This will show our mapping of which jobs are found in which PBAs

tbl = pd.DataFrame(np.zeros((len(job),len(pba_id)),dtype=bool),index = job.index,columns = pba_id.index)

Read in CBECS data

# ---------------------------------------------------------------------- #
cb = pd.read_csv('Datasets/cbecs_servers_cleaned_2012_dummy.csv')
#cb = pd.read_csv('Datasets/cbecs_servers_cleaned_2018_dummy.csv')


cb = cb[['PUBID', 'FINALWT', 'NWKER', 'PBA', 'CENDIV']]
# ---------------------------------------------------------------------- #

Create a new column to get number of workers for a given building in CBECS

cb['FINALWT_NWKER'] = cb['FINALWT']*cb['NWKER']

Do a pivot table to get the number of workers in a given PBA by census division

cb_piv = cb.pivot_table(index = 'CENDIV',columns='PBA',values = 'FINALWT_NWKER',aggfunc=np.sum)

Replace NaNs w/ 0s

cb_piv = cb_piv.replace(np.nan,0.)

There are some PBAs that are missing in cb_piv. For completeness add those PBAs back in and assign them values of 0.

no_vals = list(set(tbl.columns).difference(set(cb_piv.columns)))
for x in no_vals: cb_piv[x] = 0.


cb_piv = cb_piv.loc[:,tbl.columns]

Load in table to map state name to census division

geo = pd.read_csv('Datasets/geo_mapping.csv')
state_to_cdiv = geo.loc[:,['state abbrev','cdiv num']].set_index('state abbrev')

Add census division to occ_data

occ_data_big = df.join(state_to_cdiv,on = 'STATE')
occ_data_big_tot = df_tot.join(state_to_cdiv,on = 'STATE')

Get only koser data

m = (occ_data_big['cdiv num'] == occ_data_big['cdiv num'])

occ_data = occ_data_big[m]

Fill out truth table

for i in job.index:
    vals = job.loc[i,job.columns[1:]].dropna().astype(int).values
    tbl.loc[i,vals] = True

Use truth table to fill out a mask based on job code in occ_data

occ_mask = (occ_data.loc[:,['OCC_CODE']].join(tbl,on='OCC_CODE')).drop('OCC_CODE',axis = 1)

Use cb_piv to fill out a dataframe with number of workers in a given census division

occ_finalwt = (occ_data.loc[:,['cdiv num']].join(cb_piv,on='cdiv num')).drop('cdiv num',axis = 1)

Multiply number of works by mask and normalize

occ_mask_norm = (occ_mask * occ_finalwt).div((occ_mask * occ_finalwt).sum(axis=1),axis=0)

Now multiply by the number of workers in the OCC data

occ_by_pba = occ_mask_norm.mul(occ_data[field],axis=0)

Drop any empy rows

occ_by_pba.dropna(axis=0,how='all',inplace = True)

Flatten out the dataframe

occ_by_pba_flat  = occ_by_pba.unstack().replace(0,np.nan).dropna().reset_index().rename(columns = {'level_0':'pba','index':'occ_id',0:'nwrkrs'})

Add state, zip code, and census div to dataframe

occ = occ_by_pba_flat.join(occ_data.loc[:,['STATE','ZIP CODE','cdiv num']],on='occ_id')

Groupby zip and pba to get the number of workers in a given zip code for a given PBA

occ = occ.groupby(['STATE','ZIP CODE','pba'])['nwrkrs'].sum().reset_index()

okay, now we are going to get geographic distribution of all workers in OCC data.

m2 = (occ_data_big_tot['cdiv num'] == occ_data_big_tot['cdiv num'])

occ_data2 = occ_data_big_tot.loc[m2]

all_workers = occ_data2.loc[:,['STATE','ZIP CODE',field]]

all_workers.rename(columns = {field:'nwrkrs'},inplace = True)

Assign to PBA 1 (which currently shouldn't be included in the dataset)

all_workers['pba'] = 1

all_workers = all_workers[['STATE','ZIP CODE','pba','nwrkrs']]

Add this to the occ dataframe

occ = pd.concat([occ,all_workers],ignore_index=True)

Add census division to the dataframe

occ = occ.join(state_to_cdiv,on='STATE')

#occ = occ.sort(columns=['cdiv num','pba'])

occ = occ.sort_values(['cdiv num','pba'])

Get total number of workers for each census div and pba

occ_tot = occ.groupby(['cdiv num','pba'])[['nwrkrs']].sum()
occ_tot.rename(columns = {'nwrkrs':'nwrkrs_tot'}, inplace = True)

Add columns to occ dataframe

occ = occ.join(occ_tot,on =['cdiv num','pba'])

Divide the number of workers (nwrkrs) in each zip and pba by the total number of workers in that census divison and pba.

occ['frac'] = occ['nwrkrs']/occ['nwrkrs_tot']

occ.reset_index(drop = True, inplace = True)

Write out our csv file that has the fraction of employees in a given zip and pba relative to the number of workers that census division and pba.

# ------------------------------------------- #
occ.to_csv('Datasets/employment_by_zip_and_pba_2012.csv')
# ------------------------------------------- #

# occ[(occ['cdiv num'] == 3) & (occ['pba'] == 4)].frac.sum()

occ.head()

sorted(occ.pba.unique())
